+++ GBHFacebookImagePicker/Classes/Controller/GBHFacebookAlbumPicker.swift
@@ -13,13 +13,18 @@ protocol GBHAlbumPickerTableViewControllerDelegate: class {
     /// - parameter imageModel: model of the selected picture
     func didSelecPicturesInAlbum(imageModels: [GBHFacebookImage])
 
-    /// Failed selecte picture in album
+    /// Failed to select picture in album
     ///
     /// - parameter error: error
     func didFailSelectPictureInAlbum(error: Error?)
 }
 
 class GBHFacebookAlbumPicker: UITableViewController {
+
+    // Status bar
+    override var preferredStatusBarStyle: UIStatusBarStyle {
+        return GBHFacebookImagePicker.pickerConfig.uiConfig.statusbarStyle
+    }
 
     // MARK: - Var
 
@@ -28,9 +33,6 @@ class GBHFacebookAlbumPicker: UITableViewController {
 
     /// Album cell identifier 
     private let reuseIdentifier = "AlbumCell"
-
-    /// Loading indicator 
-    fileprivate var indicator = UIActivityIndicatorView() // Loading indicator
     
     // No album label
     fileprivate var noAlbumLabel = UILabel()
@@ -46,6 +48,18 @@ class GBHFacebookAlbumPicker: UITableViewController {
         }
     }
 
+    fileprivate lazy var loadingIndicator: UIActivityIndicatorView = {
+        let indicator = UIActivityIndicatorView(frame: CGRect(x: 0,
+                                                              y: 0,
+                                                              width: 40,
+                                                              height: 40) )
+        indicator.hidesWhenStopped = true
+        indicator.activityIndicatorViewStyle = UIActivityIndicatorViewStyle.gray
+        indicator.backgroundColor = UIColor.clear
+        indicator.color = UIColor.black
+        return indicator
+    }()
+
     // MARK: - Lifecycle
 
     override public func viewDidLoad() {
@@ -54,10 +68,10 @@ class GBHFacebookAlbumPicker: UITableViewController {
         // Prepare view
         self.prepareTableView()
         self.prepareObserver()
-        self.prepareActivityIndicator()
         self.prepareNoAlbumLabel()
 
         // Start Facebook login
+        self.startLoading()
         self.doFacebookLogin()
     }
 
@@ -67,14 +81,14 @@ class GBHFacebookAlbumPicker: UITableViewController {
     fileprivate func prepareTableView() {
         // Common init
         self.tableView.tableFooterView = UIView()
-        self.title = GBHFacebookImagePicker.pickerConfig.title
+        self.title = GBHFacebookImagePicker.pickerConfig.textConfig.localizedTitle
         self.tableView.register(GBHAlbumTableViewCell.self,
                                 forCellReuseIdentifier: self.reuseIdentifier)
         self.tableView.dataSource = self
         self.tableView.delegate = self
-        self.tableView.separatorStyle = .none
         self.tableView.cellLayoutMarginsFollowReadableWidth = false
         self.view.backgroundColor = GBHFacebookImagePicker.pickerConfig.uiConfig.backgroundColor ?? .white
+        self.tableView.backgroundView = self.loadingIndicator
 
         // Close button (on the right corner of navigation bar)
         let closeButton = UIBarButtonItem(barButtonSystemItem: .stop,
@@ -106,38 +120,12 @@ class GBHFacebookAlbumPicker: UITableViewController {
 
     // MARK: - Loading indicator
 
-    /// Create & add activity indicator to the center of view
-    fileprivate func prepareActivityIndicator() {
-        // Init 
-        self.indicator = UIActivityIndicatorView(frame:CGRect(x: 0,
-                                                              y: 0,
-                                                              width: 40,
-                                                              height: 40) )
-
-        // Style 
-        self.indicator.activityIndicatorViewStyle = UIActivityIndicatorViewStyle.gray
-
-        // Position 
-        self.indicator.center = self.view.center
-
-        // Add 
-        self.view.addSubview(indicator)
-
-        // Start loading animation 
-        self.startLoading()
-    }
-
-    /// Start loading indicator
     fileprivate func startLoading() {
-        self.indicator.startAnimating()
-        self.indicator.backgroundColor = UIColor.clear
-        self.indicator.color = UIColor.black
+        self.loadingIndicator.startAnimating()
     }
 
-    /// Stop & hide loading indicator
     fileprivate func stopLoading() {
-        self.indicator.stopAnimating()
-        self.indicator.hidesWhenStopped = true
+        self.loadingIndicator.stopAnimating()
     }
 
     // MARK: - Action
@@ -173,7 +161,7 @@ class GBHFacebookAlbumPicker: UITableViewController {
     }
 
     /// Handler for did retrieve album list
-    func didReceiveAlbum(_ sender: Notification) {
+    @objc func didReceiveAlbum(_ sender: Notification) {
         if let albums =  sender.object as? [GBHFacebookAlbum] {
             self.albums = albums
             self.stopLoading()
@@ -182,23 +170,19 @@ class GBHFacebookAlbumPicker: UITableViewController {
 
     /// Show popup with ask user_photos permission
     fileprivate func showDeniedPermissionPopup() {
-        let alertController = UIAlertController(title: NSLocalizedString("Oups",
-                                                                         comment: ""),
-                                                message: NSLocalizedString("You need to allow photo's permission.",
-                                                                           comment: ""),
+        let alertController = UIAlertController(title: GBHFacebookImagePicker.pickerConfig.textConfig.localizedOups,
+                                                message: GBHFacebookImagePicker.pickerConfig.textConfig.localizedAllowPhotoPermission,
                                                 preferredStyle: UIAlertControllerStyle.alert)
 
         // Done & Cancel button
-        let autorizeAction = UIAlertAction(title: NSLocalizedString("Allow",
-                                                                    comment: ""),
+        let autorizeAction = UIAlertAction(title: GBHFacebookImagePicker.pickerConfig.textConfig.localizedAllow,
                                            style: UIAlertActionStyle.default,
                                            handler: { _ -> Void in
                                             self.doFacebookLogin()
         })
 
         // Cancel action
-        let cancelAction = UIAlertAction(title: NSLocalizedString("Close",
-                                                                  comment: ""),
+        let cancelAction = UIAlertAction(title: GBHFacebookImagePicker.pickerConfig.textConfig.localizedClose,
                                          style: UIAlertActionStyle.cancel,
                                          handler: { (_: UIAlertAction!) -> Void in
                                             self.dismissPicker()
@@ -220,23 +204,37 @@ class GBHFacebookAlbumPicker: UITableViewController {
         return 1
     }
 
-    override public func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+    override public func tableView(_ tableView: UITableView,
+                                   numberOfRowsInSection section: Int) -> Int {
         return self.albums.count
     }
 
-    override public func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+    override public func tableView(_ tableView: UITableView,
+                                   cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        // Dequeue the cell 
         var cell = tableView.dequeueReusableCell(withIdentifier: self.reuseIdentifier,
                                                  for: indexPath) as? GBHAlbumTableViewCell
-        if cell == nil {
-            cell = GBHAlbumTableViewCell(style: .subtitle, reuseIdentifier: self.reuseIdentifier)
-        }
 
-        cell?.configure(album: albums[indexPath.row])
+        if cell == nil {
+            cell = GBHAlbumTableViewCell(style: .subtitle,
+                                         reuseIdentifier: self.reuseIdentifier)
+        }
 
         return cell!
     }
 
-    override public func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
+    override func tableView(_ tableView: UITableView,
+                            willDisplay cell: UITableViewCell,
+                            forRowAt indexPath: IndexPath) {
+
+        if let cell = cell as? GBHAlbumTableViewCell {
+            // Configure the cell 
+            cell.configure(album: albums[indexPath.row])
+        }
+    }
+
+    override public func tableView(_ tableView: UITableView,
+                                   heightForRowAt indexPath: IndexPath) -> CGFloat {
         return 90
     }
 
@@ -254,6 +252,10 @@ class GBHFacebookAlbumPicker: UITableViewController {
     /// Dismiss the picker 
     func dismissPicker() {
         DispatchQueue.main.async {
+            // Reset flag
+            GBHFacebookManager.shared.reset()
+
+            // Dismiss and call delegate 
             self.dismiss(animated: true, completion: {
                 self.delegate?.facebookImagePickerDismissed()
             })
@@ -274,52 +276,38 @@ extension GBHFacebookAlbumPicker: GBHAlbumPickerTableViewControllerDelegate {
         var errorModels = [GBHFacebookImage]()
         var errors = [Error?]()
 
-        func callback() {
-            // Check if all image finished loading 
-            if successModels.count + errors.count == imageModels.count {
-                // Call success delegate 
-                self.delegate?.facebookImagePicker(
-                    imagePicker: self,
-                    successImageModels: successModels,
-                    errorImageModels: errorModels,
-                    errors: errors
-                )
-
-                // Dismiss picker 
-                self.dismissPicker()
-            }
-        }
+        let downloadGroup = DispatchGroup()
 
         for imageModel in imageModels {
-            if let url = imageModel.fullSizeUrl, let imageUrl = URL(string: url) {
-                // Start url loading
-                URLSession.shared.dataTask(with: imageUrl as URL) { data, response, error in
-                    guard let data = data, error == nil else {
-                        errorModels.append(imageModel)
-                        errors.append(error)
-                        callback()
-                        return
-                    }
-                    if let httpResponse = response as? HTTPURLResponse, httpResponse.statusCode != 200 {
-                        errorModels.append(imageModel)
-                        errors.append(error)
-                        callback()
-                        return
-                    }
+            downloadGroup.enter()
 
-                    // Set the image
-                    imageModel.image = UIImage(data: data)
-
+            // Download the image from the full size url
+            imageModel.download(completion: { (error) in
+                if error != nil {
+                    // Error case
+                    errors.append(error)
+                    errorModels.append(imageModel)
+                } else {
+                    // Success case 
                     successModels.append(imageModel)
-                    callback()
-                    }.resume()
-            } else {
-                errorModels.append(imageModel)
-                errors.append(nil)
-                callback()
-            }
+                }
+
+                downloadGroup.leave()
+            })
         }
 
+        downloadGroup.notify(queue: .main) {
+            // Call success delegate
+            self.delegate?.facebookImagePicker(
+                imagePicker: self,
+                successImageModels: successModels,
+                errorImageModels: errorModels,
+                errors: errors
+            )
+
+            // Dismiss picker
+            self.dismissPicker()
+        }
     }
 
     /// Performed when an error occured