+++ GBHFacebookImagePicker/Classes/Model/GBHFacebookImage.swift
@@ -5,6 +5,11 @@
 //  Created by Florian Gabach on 29/09/2016.
 //  Copyright (c) 2016 Florian Gabach <contact@floriangabach.fr>
 
+public enum ImageSize {
+    case normal
+    case full
+}
+
 public class GBHFacebookImage {
 
     // MARK: - Var
@@ -33,5 +38,33 @@ public class GBHFacebookImage {
         self.imageId = imgId
         self.normalSizeUrl = picture
         self.fullSizeUrl = source
+    }
+
+    // MARK: - Download
+
+    /// Download the image
+    ///
+    /// - Parameter completion: completion handler with optional error 
+    internal func download(completion: @escaping (Error?) -> Void) {
+        guard let stringUrl = self.fullSizeUrl,
+            let url = URL(string: stringUrl) else {
+                completion(DownloadError.invalidUrl)
+                return
+        }
+
+        URLSession.shared.dataTask(with: url) { data, response, error in
+            guard error == nil,
+                let data = data,
+                let httpResponse = response as? HTTPURLResponse,
+                httpResponse.statusCode == 200 else {
+                    completion(DownloadError.downloadError)
+                    return
+            }
+
+            // Set the image
+            self.image = UIImage(data: data)
+            completion(nil)
+            }
+            .resume()
     }
 }